Fixes some small bugs and updates Permission access to allow users of OSU to be able to use the bot as expected.

+++ src/main/java/com/crazy/scientist/crazyjavascientist/commands/TaskManager.java
@@ -34,43 +34,45 @@ public class TaskManager {
     UserTaskTableI userTaskTable;
 
 
-    public void createNewUserList(@Nonnull SlashCommandInteraction event) {
+    public void createNewUserList(boolean hasPermission, @Nonnull SlashCommandInteraction event) {
 
-        UserTaskTable taskTable = new UserTaskTable();
+        if (hasPermission) {
+
+            UserTaskTable taskTable = new UserTaskTable();
 
 
-        int numOfTaskLists = userTaskTable.getUserTaskListCount(event.getUser().getName());
+            int numOfTaskLists = userTaskTable.getUserTaskListCount(event.getUser().getName());
 
 
-        if (numOfTaskLists == 5) {
+            if (numOfTaskLists == 5) {
 
-            event.reply("You have hit the limit of Lists allowed, The limit is 5. Please Remove a list before adding a new one").queue();
-        } else {
+                event.reply("You have hit the limit of Lists allowed, The limit is 5. Please Remove a list before adding a new one").queue();
+            } else {
 
-            if (event.getOptions().isEmpty()) {
+                if (event.getOptions().isEmpty()) {
 
-                TextInput task_list_title = TextInput.create("task-list-title","Task List Title", TextInputStyle.SHORT).build();
+                    TextInput task_list_title = TextInput.create("task-list-title", "Task List Title", TextInputStyle.SHORT).build();
 
-                TextInput task_list_description = TextInput.create("task-list-description","First Task's Description",TextInputStyle.PARAGRAPH).build();
+                    TextInput task_list_description = TextInput.create("task-list-description", "First Task's Description", TextInputStyle.PARAGRAPH).build();
 
 //                TextInput task_list_status = TextInput.create("task-list-status","First Task's Status",TextInputStyle.UNKNOWN).build();
 
 
-                TextInput task_list_comments = TextInput.create("task-list-comment","First Task's Comments",TextInputStyle.PARAGRAPH).build();
+                    TextInput task_list_comments = TextInput.create("task-list-comment", "First Task's Comments", TextInputStyle.PARAGRAPH).build();
 
-                TextInput statusMenu = TextInput.create("status-menu","First Task's Status",TextInputStyle.SHORT).build();
-
-
-                Modal taskListModal = Modal.create("create-task-list-modal", "Create A Task List!")
-                        .addActionRows(ActionRow.of(task_list_title), ActionRow.of(task_list_description),ActionRow.of(statusMenu),ActionRow.of(task_list_comments))
-                        .build();
+                    TextInput statusMenu = TextInput.create("status-menu", "First Task's Status", TextInputStyle.SHORT).build();
 
 
-                event.replyModal(taskListModal).queue();
+                    Modal taskListModal = Modal.create("create-task-list-modal", "Create A Task List!")
+                            .addActionRows(ActionRow.of(task_list_title), ActionRow.of(task_list_description), ActionRow.of(statusMenu), ActionRow.of(task_list_comments))
+                            .build();
+
+
+                    event.replyModal(taskListModal).queue();
 
                 /*taskTable.setTaskTitle(event.getOption("title").getAsString());
                 taskTable.setUsername(event.getUser().getName());*/
-            } else {
+                } else {
 
               /*  taskTable.setUsername(event.getUser().getName());
                 taskTable.setTaskTitle(event.getOption("title").getAsString());
@@ -78,43 +80,51 @@ public class TaskManager {
                 taskTable.setTaskStatus(event.getOption("status").getAsString());
                 taskTable.setTaskComments(event.getOption("comments").getAsString());*/
 
-                TextInput task_list_title = TextInput.create("task-list-title","Task List Title", TextInputStyle.SHORT).build();
+                    TextInput task_list_title = TextInput.create("task-list-title", "Task List Title", TextInputStyle.SHORT).build();
 
-                TextInput task_list_description = TextInput.create("task-list-description","First Task's Description",TextInputStyle.PARAGRAPH).build();
+                    TextInput task_list_description = TextInput.create("task-list-description", "First Task's Description", TextInputStyle.PARAGRAPH).build();
 
 //                TextInput task_list_status = TextInput.create("task-list-status","First Task's Status",TextInputStyle.UNKNOWN).build();
 
 
-                TextInput task_list_comments = TextInput.create("task-list-comment","First Task's Comments",TextInputStyle.PARAGRAPH).build();
+                    TextInput task_list_comments = TextInput.create("task-list-comment", "First Task's Comments", TextInputStyle.PARAGRAPH).build();
 
-                SelectMenu statusMenu = SelectMenu.create("status-menu").addOption("status","Status","sets the status for the task at hand").build();
+                    SelectMenu statusMenu = SelectMenu.create("status-menu").addOption("status", "Status", "sets the status for the task at hand").build();
 
-                Modal taskListModal = Modal.create("create-task-list-modal", "Create A Task List!")
-                        .addActionRows(ActionRow.of(task_list_title), ActionRow.of(task_list_description),ActionRow.of(statusMenu),ActionRow.of(task_list_comments))
-                        .build();
-                event.replyModal(taskListModal).queue();
+                    Modal taskListModal = Modal.create("create-task-list-modal", "Create A Task List!")
+                            .addActionRows(ActionRow.of(task_list_title), ActionRow.of(task_list_description), ActionRow.of(statusMenu), ActionRow.of(task_list_comments))
+                            .build();
+                    event.replyModal(taskListModal).queue();
 
+                }
+
+                userTaskTable.save(taskTable);
+                event.reply("List Created Successfully").queue();
             }
 
-            userTaskTable.save(taskTable);
-            event.reply("List Created Successfully").queue();
+        } else {
+            event.replyFormat("You are not allowed to use slash commands%n Please reach out to  %s  and he can allow you to use commands ", event.getJDA().getUserById(416342612484554752L).getName()).queue();
         }
 
 
     }
 
-    public void deleteUserTaskListByTitle(@Nonnull SlashCommandInteraction event){
+    public void deleteUserTaskListByTitle(boolean hasPermission,@Nonnull SlashCommandInteraction event) {
 
 
-        if(userTaskTable.checkIfUserTaskListExists(event.getOption("title").getAsString(),event.getUser().getName()).isEmpty()){
+        if(hasPermission) {
+            if (userTaskTable.checkIfUserTaskListExists(event.getOption("title").getAsString(), event.getUser().getName()).isEmpty()) {
 
-            event.replyFormat("No User Task Lists with the name %s found",event.getOption("title").getAsString()).queue();
+                event.replyFormat("No User Task Lists with the name %s found", event.getOption("title").getAsString()).queue();
+            } else {
+
+                userTaskTable.deleteTaskList(event.getOption("title").getAsString(), event.getUser().getName());
+
+                event.replyFormat("User Task List \"%s\" has been deleted successfully!", event.getOption("title").getAsString()).queue();
+
+            }
         }else{
-
-            userTaskTable.deleteTaskList(event.getOption("title").getAsString(),event.getUser().getName());
-
-            event.replyFormat("User Task List \"%s\" has been deleted successfully!",event.getOption("title").getAsString()).queue();
-
+            event.replyFormat("You are not allowed to use slash commands%n Please reach out to  %s  and he can allow you to use commands ", event.getJDA().getUserById(416342612484554752L).getName()).queue();
         }
